reorganize traits to allow proper wasm compilation
