+++ shell.nix
@@ -6,10 +6,16 @@ in
   with pkgs;
 
   pkgs.mkShell {
+    nativeBuildInputs = [
+      clang_13
+    ];
+
     buildInputs = [
       rustPlatform.rust.cargo
+      openssl
+      pkg-config
+
       git
-      stdenv.cc.cc.lib
       (python38.withPackages(ps: with ps; [ virtualenv tox setuptools ]))
       (python39.withPackages(ps: with ps; [ virtualenv setuptools ]))
       (python37.withPackages(ps: with ps; [ virtualenv setuptools ]))
@@ -25,11 +31,14 @@ in
       cargo-watch
       cargo-limit
       cargo-udeps
+
+      llvmPackages_13.libclang
+      llvmPackages_13.libcxxClang
     ];
 
-    shellHook = ''
-       # workaround for https://github.com/NixOS/nixpkgs/blob/48dfc9fa97d762bce28cc8372a2dd3805d14c633/doc/languages-frameworks/python.section.md#python-setuppy-bdist_wheel-cannot-create-whl
-       export SOURCE_DATE_EPOCH=315532800 # 1980
-       export LD_LIBRARY_PATH="${stdenv.cc.cc.lib}/lib64:$LD_LIBRARY_PATH";
-    '';
+    BINDGEN_EXTRA_CLANG_ARGS = "-isystem ${llvmPackages_13.libclang.lib}/lib/clang/${lib.getVersion clang}/include";
+    LIBCLANG_PATH = "${llvmPackages_13.libclang.lib}/lib";
+
+    # workaround for https://github.com/NixOS/nixpkgs/blob/48dfc9fa97d762bce28cc8372a2dd3805d14c633/doc/languages-frameworks/python.section.md#python-setuppy-bdist_wheel-cannot-create-whl
+    SOURCE_DATE_EPOCH = 315532800; # 1980
   }
