+++ src/core/Cargo.toml
@@ -80,3 +80,6 @@ features = ["console", "File"]
 
 [target.'cfg(all(target_arch = "wasm32", target_vendor="unknown"))'.dev-dependencies]
 wasm-bindgen-test = "0.3.0"
+
+### These crates don't compile on wasm
+[target.'cfg(not(all(target_arch = "wasm32", target_vendor="unknown")))'.dependencies]
