+++ src/core/src/index/mod.rs
@@ -7,8 +7,6 @@ pub mod bigsi;
 pub mod linear;
 pub mod sbt;
 
-pub mod storage;
-
 pub mod search;
 
 use std::ops::Deref;
@@ -19,12 +17,14 @@ use once_cell::sync::OnceCell;
 use serde::{Deserialize, Serialize};
 use typed_builder::TypedBuilder;
 
+use crate::errors::ReadDataError;
 use crate::index::sbt::{Node, SBT};
 use crate::index::search::{search_minhashes, search_minhashes_containment};
-use crate::index::storage::{ReadData, ReadDataError, Storage};
-use crate::signature::{Signature, SigsTrait};
+use crate::prelude::*;
+use crate::signature::SigsTrait;
 use crate::sketch::nodegraph::Nodegraph;
 use crate::sketch::Sketch;
+use crate::storage::Storage;
 use crate::Error;
 
 pub type MHBT = SBT<Node<Nodegraph>, Signature>;
@@ -101,12 +101,6 @@ pub trait Index<'a> {
     /*
     fn iter_signatures(&self) -> Self::SignatureIterator;
     */
-}
-
-// TODO: split into two traits, Similarity and Containment?
-pub trait Comparable<O> {
-    fn similarity(&self, other: &O) -> f64;
-    fn containment(&self, other: &O) -> f64;
 }
 
 impl<'a, N, L> Comparable<L> for &'a N
