+++ src/core/src/lib.rs
@@ -22,12 +22,13 @@
 pub mod errors;
 pub use errors::SourmashError as Error;
 
-pub mod cmd;
+pub mod prelude;
 
-pub mod index;
+pub mod cmd;
 
 pub mod signature;
 pub mod sketch;
+pub mod storage;
 
 pub mod encodings;
 
@@ -42,6 +43,7 @@ cfg_if! {
         pub mod wasm;
     } else {
         pub mod ffi;
+        pub mod index;
     }
 }
